51: implemented credit card testing

+++ server/controllers/creditCardController.js
@@ -2,6 +2,7 @@
 const xssFilters = require('xss-filters');
 const CreditCard = require('../models/CreditCard');
 const UserRole = require('../models/UserRole');
+const User = require('../models/User');
 
 module.exports = {
   /**
@@ -60,39 +61,53 @@ module.exports = {
 
   async showCreditCard(req, res) {
     if (req.userRole) {
-      try {
-        const currentURL = `${req.protocol}://${req.get('host')}${
-          req.originalUrl
-        }`;
-        let card = await CreditCard.findById(req.params.cardId).exec();
-        card = JSON.parse(JSON.stringify(card));
+      User.findById(req.userId, async (uerr, doc) => {
+        if (uerr) {
+          res.status(400).json({ message: uerr });
+        } else {
+          // Check if creditCardId matches or user is admin
+          if (
+            doc.creditCardId === req.params.cardId ||
+            req.userRole === UserRole.ADMIN
+          ) {
+            try {
+              const currentURL = `${req.protocol}://${req.get('host')}${
+                req.originalUrl
+              }`;
+              let card = await CreditCard.findById(req.params.cardId).exec();
+              card = JSON.parse(JSON.stringify(card));
 
-        // Add HATEOAS links to credit card
-        card.links = [
-          {
-            rel: 'self',
-            method: 'GET',
-            href: currentURL,
-            types: ['application/json'],
-          },
-          {
-            rel: 'self',
-            method: 'DELETE',
-            href: currentURL,
-            types: [],
-          },
-          {
-            rel: 'self',
-            method: 'PUT',
-            href: currentURL,
-            types: ['application/json'],
-          },
-        ];
-        res.status(200).json(card);
-      } catch (err) {
-        res.status(404).json({ message: err });
-        console.log('Caught an error: ', err);
-      }
+              // Add HATEOAS links to credit card
+              card.links = [
+                {
+                  rel: 'self',
+                  method: 'GET',
+                  href: currentURL,
+                  types: ['application/json'],
+                },
+                {
+                  rel: 'self',
+                  method: 'DELETE',
+                  href: currentURL,
+                  types: [],
+                },
+                {
+                  rel: 'self',
+                  method: 'PUT',
+                  href: currentURL,
+                  types: ['application/json'],
+                },
+              ];
+              res.status(200).json(card);
+            } catch (err) {
+              res.status(404).json({ message: err });
+              console.log('Caught an error: ', err);
+            }
+          } else {
+            res.status(403).json({ message: 'Forbidden' });
+          }
+        }
+      });
     } else {
       res.status(401).json({ message: 'Unauthorized' });
     }
@@ -105,31 +120,41 @@ module.exports = {
    */
   async addCreditCard(req, res) {
     if (req.userRole) {
-      try {
-        const { number, CVC, ownerName } = req.body;
-        let card = await CreditCard.findOne({ number }).exec();
+      User.findById(req.userId, async (uerr, doc) => {
+        if (uerr) {
+          res.status(400).json({ message: uerr });
+        } else {
+          try {
+            const { number, CVC, ownerName } = req.body;
+            let card = await CreditCard.findOne({ number }).exec();
 
-        if (card) {
-          const errorMessage = 'Card already registered in DB!';
+            if (card) {
+              const errorMessage = 'Card already registered in DB!';
 
-          if (!req.is('json')) {
-            res.status(400).json({ message: errorMessage });
+              if (!req.is('json')) {
+                res.status(400).json({ message: errorMessage });
+              }
+            }
+
+            card = new CreditCard({
+              number: xssFilters.inHTMLData(number),
+              CVC: xssFilters.inHTMLData(CVC),
+              ownerName: xssFilters.inHTMLData(ownerName),
+            });
+
+            const newCard = await card.save();
+            await User.findByIdAndUpdate(
+              { _id: doc._id },
+              { creditCardId: newCard._id }
+            );
+            await res.status(200).json(newCard);
+            console.log('New credit card added.');
+          } catch (err) {
+            res.json({ message: err });
+            console.log('Caught an error: ', err);
           }
         }
-
-        card = new CreditCard({
-          number: xssFilters.inHTMLData(number),
-          CVC: xssFilters.inHTMLData(CVC),
-          ownerName: xssFilters.inHTMLData(ownerName),
-        });
-
-        const newCard = await card.save();
-        await res.status(201).json(newCard);
-        console.log('New credit card added.');
-      } catch (err) {
-        res.json({ message: err });
-        console.log('Caught an error: ', err);
-      }
+      });
     } else {
       res.status(401).json({ message: 'Unauthorized' });
     }
@@ -142,15 +167,33 @@ module.exports = {
    */
   async removeCreditCard(req, res) {
     if (req.userRole) {
-      try {
-        const ID = req.params.cardId;
-        const deletedCard = await CreditCard.deleteOne({ _id: `${ID}` });
-        res.status(200).json(deletedCard);
-        console.log('Successfully removed credit card.');
-      } catch (err) {
-        res.status(400).json({ message: err });
-        console.log('Caught an error: ', err);
-      }
+      User.findById(req.userId, async (uerr, doc) => {
+        if (uerr) {
+          res.status(400).json({ message: uerr });
+        } else {
+          // Check if creditCardId matches or user is admin
+          if (
+            doc.creditCardId === req.params.cardId ||
+            req.userRole === UserRole.ADMIN
+          ) {
+            try {
+              const ID = req.params.cardId;
+              const deletedCard = await CreditCard.deleteOne({ _id: `${ID}` });
+              await User.findByIdAndUpdate(
+                { _id: doc._id },
+                { bankAccountId: 'undefined' }
+              );
+              res.status(200).json(deletedCard);
+              console.log('Successfully removed credit card.');
+            } catch (err) {
+              res.status(400).json({ message: err });
+              console.log('Caught an error: ', err);
+            }
+          } else {
+            res.status(403).json({ message: 'Forbidden' });
+          }
+        }
+      });
     } else {
       res.status(401).json({ message: 'Unauthorized' });
     }
@@ -163,18 +206,32 @@ module.exports = {
    */
   async updateCreditCard(req, res) {
     if (req.userRole) {
-      try {
-        const ID = req.params.cardId;
-        const { number, CVC, ownerName } = req.body;
-        const updatedCard = await CreditCard.findOne({ _id: `${ID}` });
-        updatedCard.number = xssFilters.inHTMLData(number);
-        updatedCard.CVC = xssFilters.inHTMLData(CVC);
-        updatedCard.ownerName = xssFilters.inHTMLData(ownerName);
-        await updatedCard.save();
-        res.status(200).json(updatedCard);
-      } catch (err) {
-        res.status(400).json({ message: err });
-      }
+      User.findById(req.userId, async (uerr, doc) => {
+        if (uerr) {
+          res.status(400).json({ message: uerr });
+        } else {
+          // Check if creditCardId matches or user is admin
+          if (
+            doc.creditCardId === req.params.cardId ||
+            req.userRole === UserRole.ADMIN
+          ) {
+            try {
+              const ID = req.params.cardId;
+              const { number, CVC, ownerName } = req.body;
+              const updatedCard = await CreditCard.findOne({ _id: `${ID}` });
+              updatedCard.number = xssFilters.inHTMLData(number);
+              updatedCard.CVC = xssFilters.inHTMLData(CVC);
+              updatedCard.ownerName = xssFilters.inHTMLData(ownerName);
+              await updatedCard.save();
+              res.status(200).json(updatedCard);
+            } catch (err) {
+              res.status(400).json({ message: err });
+            }
+          } else {
+            res.status(403).json({ message: 'Forbidden' });
+          }
+        }
+      });
     } else {
       res.status(401).json({ message: 'Unauthorized' });
     }
